CJA-026: Merge tracing-common into core cja crate
Move the tracing-common crate's code directly into cja::setup,
replacing the glob re-export with the full implementation. This
eliminates an unnecessary workspace member while preserving the
identical public API (setup_sentry, setup_tracing, EyesShutdownHandle).

- Add tracing-common's dependencies to cja/Cargo.toml
- Replace setup.rs glob re-export with full source
- Remove crates/tracing-common/ directory
- Update CLAUDE.md architecture section
