+++ crates/cja/Cargo.toml
@@ -34,8 +34,23 @@ listenfd = "1.0.2"
 cron = "0.12.1"
 chrono-tz = "0.10.0"
 
-tracing-common = { path = "../tracing-common" }
 color-eyre = "0.6.3"
+eyes-subscriber = "0.1.4"
+opentelemetry = { workspace = true }
+opentelemetry_sdk = { workspace = true }
+opentelemetry-otlp = { workspace = true }
+sentry = { version = "0.31.5", default-features = false, features = [
+  "rustls",
+  "backtrace",
+  "contexts",
+  "panic",
+  "tower",
+  "reqwest",
+] }
+sentry-tracing = "0.31.8"
+tracing-opentelemetry = { workspace = true }
+tracing-subscriber = { workspace = true }
+tracing-tree = { workspace = true }
 serde_urlencoded = { version = "0.7", optional = true }
 maud.workspace = true
 tokio-util.workspace = true
