--- crates/tracing-common/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "tracing-common"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-sentry = { version = "0.31.5", default-features = false, features = [
-  "rustls",
-  "backtrace",
-  "contexts",
-  "panic",
-  "tower",
-  "reqwest",
-] }
-sentry-tracing = "0.31.8"
-tracing = { workspace = true }
-tracing-opentelemetry = { workspace = true }
-tracing-subscriber = { workspace = true }
-tracing-tree = { workspace = true }
-opentelemetry = { workspace = true }
-opentelemetry_sdk = { workspace = true }
-opentelemetry-otlp = { workspace = true }
-
-color-eyre = "0.6.3"
-eyes-subscriber = "0.1.4"
-uuid = { workspace = true }
